+++ core/src/storage/IOWriter.h
@@ -11,6 +11,7 @@
 
 #pragma once
 
+#include <memory>
 #include <string>
 
 namespace milvus {
@@ -18,9 +19,8 @@ namespace storage {
 
 class IOWriter {
  public:
-    explicit IOWriter(const std::string& name) : name_(name), len_(0) {
-    }
-    ~IOWriter() = default;
+    virtual void
+    open(const std::string& name) = 0;
 
     virtual void
     write(void* ptr, size_t size) = 0;
@@ -28,10 +28,11 @@ class IOWriter {
     virtual size_t
     length() = 0;
 
- public:
-    std::string name_;
-    size_t len_;
+    virtual void
+    close() = 0;
 };
+
+using IOWriterPtr = std::shared_ptr<IOWriter>;
 
 }  // namespace storage
 }  // namespace milvus